* Added the possibility to start the software in a subpath.
* Added the default arch presets that made the software think that the Pavlov servers were not running and it could not be started.
* Rcon commands that succeed are no longer displayed in a dialog, but are just notifications
* Shack support was added to matches and much more, but the whole matchmaking system is still in testing phase
* Added Shack support with one mappool per ssh server
* GodMode added
* Added custom commands
* Added ability to give commands to all players at once
* Fixed the behavior when removing the server from the hard drive

+++ PavlovRconWebserver/Controllers/PavlovServerController.cs
@@ -24,6 +24,7 @@ namespace PavlovRconWebserver.Controllers
         private readonly PavlovServerService _pavlovServerService;
         private readonly ServerSelectedMapService _serverSelectedMapService;
         private readonly ServerSelectedModsService _serverSelectedModsService;
+        private readonly SshServerSerivce _sshServerSerivce;
         private readonly SshServerSerivce _service;
         private readonly SteamIdentityService _steamIdentityService;
         private readonly UserService _userservice;
@@ -37,6 +38,7 @@ namespace PavlovRconWebserver.Controllers
             PavlovServerService pavlovServerService,
             ServerSelectedMapService serverSelectedMapService,
             MapsService mapsService,
+            SshServerSerivce sshServerSerivce,
             ServerSelectedWhitelistService whitelistService,
             SteamIdentityStatsServerService steamIdentityStatsServerService,
             ServerSelectedModsService serverSelectedModsService,
@@ -53,6 +55,7 @@ namespace PavlovRconWebserver.Controllers
             _mapsService = mapsService;
             _whitelistService = whitelistService;
             _steamIdentityService = steamIdentityService;
+            _sshServerSerivce = sshServerSerivce;
             _serverSelectedModsService = serverSelectedModsService;
             UserManager = userManager;
         }
@@ -69,6 +72,9 @@ namespace PavlovRconWebserver.Controllers
             var server = new PavlovServer();
             if (serverId != 0) server = await _pavlovServerService.FindOne(serverId);
 
+            if (server.SshServer == null)
+                server.SshServer = await _sshServerSerivce.FindOne(sshServerId);
+            
             var viewModel = new PavlovServerViewModel();
             viewModel = viewModel.fromPavlovServer(server, sshServerId);
 
@@ -101,10 +107,15 @@ namespace PavlovRconWebserver.Controllers
                 return Forbid();
             var serverSelectedMap = new List<ServerSelectedMap>();
             var server = await _pavlovServerService.FindOne(serverId);
+            
             serverSelectedMap = (await _serverSelectedMapService.FindAllFrom(server)).ToList();
 
             var tmp = await _mapsService.FindAll();
 
+            if (server.Shack)
+                tmp = tmp.Where(x => x.Shack && x.ShackSshServerId == server.SshServer.Id).ToArray();
+            else
+                tmp = tmp.Where(x => !x.Shack).ToArray();
             var viewModel = new SelectedServerMapsViewModel
             {
                 AllMaps = tmp.ToList(),
@@ -230,7 +241,11 @@ namespace PavlovRconWebserver.Controllers
                 await _userservice.getUserFromCp(HttpContext.User), server.Id, _service, _pavlovServerService))
                 return Forbid();
             if (remove)
-                await _service.RemovePavlovServerFromDisk(server);
+            {
+                var isDevelopment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Development";
+                if(!isDevelopment)
+                    await _service.RemovePavlovServerFromDisk(server);
+            }
             ModelState.AddModelError("Id", error
             );
             return await EditServer(server);
@@ -295,6 +310,7 @@ namespace PavlovRconWebserver.Controllers
                 return Forbid();
             try
             {
+                viewModel.SshServer = await _sshServerSerivce.FindOne(viewModel.sshServerId);
                 var result = await _service.RemovePavlovServerFromDisk(viewModel);
                 if (result.Value != null) return BadRequest(result.Value);
 
@@ -380,7 +396,7 @@ namespace PavlovRconWebserver.Controllers
                 var connectionResult = await RconStatic.GetServerLog(server, _pavlovServerService);
                 if (connectionResult.Success)
                 {
-                    var replace = connectionResult.answer.Replace(Environment.NewLine, "<br/>");
+                    var replace = connectionResult.answer.Replace("\n", "<br/>");
                     return View("ServerLogs",replace);
                 }
                 else
@@ -477,11 +493,15 @@ namespace PavlovRconWebserver.Controllers
                 await _userservice.getUserFromCp(HttpContext.User), serverId, _service, _pavlovServerService))
                 return Forbid();
             var server = await _pavlovServerService.FindOne(serverId);
+                
             var steamIds = (await _steamIdentityService.FindAll()).ToArray();
             var selectedSteamIds = (await _whitelistService.FindAllFrom(server)).ToArray();
+            if (server.Shack)
+                steamIds = steamIds.Where(x => !string.IsNullOrEmpty(x.OculusId)).ToArray();
             //service
             var model = new PavlovServerWhitelistViewModel
             {
+                Shack = server.Shack,
                 steamIds = selectedSteamIds.Select(x => x.SteamIdentityId).ToList(),
                 pavlovServerId = server.Id
             };
